+++ organization-service-rust-crate/src/lib.rs
@@ -3,6 +3,7 @@ pub mod controllers;
 pub mod data;
 pub mod initializers;
 pub mod models;
+pub mod openapi;
 pub mod streaming;
 pub mod tasks;
 pub mod workers;